Adds a logo shortcode
The logo shortcode will switch between light or dark variants of a an organizations logo based on the current theme. If we only have the light variant, then we'll use a white background instead when the dark theme is active.

Also adds a custom created dark logo for ecocommons based on their guidance - though they have no such official branding.