Replace stackblitz with code snippet
There were issues getting the embedded iframe to run correctly.

I have deemed that it would be a maintanence burden to maintain CORS
headers for this project when they are only needed for this one blog
page.
Especially when it did not provide much more benefit over an inline code
block and a link to the stack blitz workspace.